Updated Notebooks
Updated notbooks, removed subfolder
